pyproject.toml
[tool.poetry]
name = "jawafdehi-agents"
version = "0.1.0"
description = "Flyte-based agent orchestration service for Jawafdehi workflows"
authors = ["Jawafdehi Inquiry <inquiry@jawafdehi.org>"]
readme = "README.md"
packages = [{include = "jawafdehi_agents", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.12.5"
pydantic-settings = "^2.11.0"
typer = "^0.16.0"
flyte = {version = ">=2.0.0b1,<3.0.0", allow-prereleases = true}
httpx = "^0.28.1"
markitdown = {extras = ["docx", "pdf", "pptx", "xlsx"], version = "^0.1.5"}
openai = "^1.79.0"

[tool.poetry.group.dev.dependencies]
hypothesis = "^6.148.2"
pytest = "^9.0.1"
pytest-asyncio = "^1.3.0"
black = "^26.3.0"
ruff = "^0.15.5"
opentelemetry-sdk = "^1.36.0"

[tool.poetry.scripts]
jawaf = "jawafdehi_agents.cli:main"

[tool.pytest.ini_options]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
